+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### 0.4.1
 
-- Add `geth-compat` feature that makes keystore creation/handling
-  compatible with geth [#12](https://github.com/roynalnaruto/eth-keystore-rs/pull/12)
+- Add `gocore-compat` feature that makes keystore creation/handling
+  compatible with gocore [#12](https://github.com/roynalnaruto/eth-keystore-rs/pull/12)
 
 ### 0.4.0